اتمام بخش مرجوعی و انتقال بین انبار ها واتمام بخش انبار داری

+++ src/main/java/com/example/Sanjaghak/Service/OrderItemService.java
@@ -1,14 +1,13 @@
 package com.example.Sanjaghak.Service;
 
-import com.example.Sanjaghak.Repository.OrderItemRepository;
-import com.example.Sanjaghak.Repository.OrdersRepository;
-import com.example.Sanjaghak.Repository.ProductRepository;
-import com.example.Sanjaghak.Repository.ProductVariantsRepository;
+import com.example.Sanjaghak.Enum.OrderStatus;
+import com.example.Sanjaghak.Repository.*;
 import com.example.Sanjaghak.Specification.OrderItemSpecifications;
 import com.example.Sanjaghak.Specification.OrderSpecifications;
 import com.example.Sanjaghak.model.*;
 import com.example.Sanjaghak.security.jwt.JwtUtil;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.persistence.criteria.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +16,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -34,21 +36,39 @@ public class OrderItemService {
     @Autowired
     private ProductVariantsRepository productVariantsRepository;
 
-    public OrderItem addOrderItem(OrderItem orderItem, UUID orderId, UUID variantId, String token) {
+    @Autowired
+    private CustomerRepository customerRepository;
+
+    @Autowired
+    private DiscountRepository discountRepository;
+
+    @Autowired
+    private InventoryStockRepository inventoryStockRepository;
+
+
+    public OrderItem addOrderItem(OrderItem orderItem,UUID variantId, String token) {
 
         UUID userId = UUID.fromString(JwtUtil.extractUserId(token));
         String role = JwtUtil.extractUserRole(token);
 
-        if(!ordersRepository.existsById(orderId)) {
+        Customer customer = customerRepository.findByUserId_Id(userId)
+                .orElseThrow(() -> new RuntimeException("خریدار پیدا نشد"));
+
+        Optional<Orders> order = ordersRepository.findByCustomerIdAndOrderStatus(customer, OrderStatus.pending);
+
+        Orders orders = ordersRepository.findById(order.get().getOrderId()).orElseThrow(() -> new RuntimeException("سفارش مورد نظر پیدا نشد!"));
+
+        if (order.isEmpty()) {
+            throw new RuntimeException("هیچ سفارشی با وضعیت PENDING برای این خریدار یافت نشد");
+        }
+
+        if(!ordersRepository.existsById(order.get().getOrderId())) {
             throw new IllegalArgumentException("سفارش مورد نظر یافت نشد");
         }
 
         if(!productVariantsRepository.existsById(variantId)) {
             throw new IllegalArgumentException("کالای مورد نظر یافت نشد");
         }
-
-        Orders orders = ordersRepository.findById(orderId)
-                .orElseThrow(() -> new RuntimeException("سفارش پیدا نشد"));
 
         ProductVariants products = productVariantsRepository.findById(variantId)
                 .orElseThrow(() -> new RuntimeException("محصول پیدا نشد"));
@@ -57,82 +77,60 @@ public class OrderItemService {
             throw new IllegalArgumentException("تعداد کالا نباید صفر باشد");
         }
 
-        if(orderItem.getUnitPrice() == null){
-            throw new IllegalArgumentException("قیمت کالا نباید خالی باشد");
+        if(orderItemRepository.existsByOrderIdAndVariantId(order.get(),products)){
+            throw new IllegalArgumentException("از این کالا قبلا به سبد خرید شما اضافه شده است");
         }
 
-        if(orderItem.getSubTotal() == null){
-            throw new IllegalArgumentException("مجموع قیمت کالا نباید خالی باشد");
+        Integer stock = inventoryStockRepository.getTotalStockByVariantId(variantId);
+
+        if(stock < orderItem.getQuantity()){
+            throw new IllegalArgumentException("موجودی ناکافی است !");
         }
 
-        if(orderItem.getTotalAmount() == null){
-            throw new IllegalArgumentException("قیمت نهایی نباید خالی باشد");
-        }
-
-        orderItem.setOrderId(orders);
+        orderItem.setOrderId(order.get());
         orderItem.setVariantId(products);
         orderItem.setCreatedAt(LocalDateTime.now());
-
         return orderItemRepository.save(orderItem);
     }
 
-    public OrderItem updateOrderItem(UUID orderItemId,OrderItem updatedOrderItem, UUID orderId,UUID variantId, String token) {
+    public OrderItem updateOrderItem(UUID orderItemId,OrderItem updatedOrderItem, String token) {
 
         UUID userId = UUID.fromString(JwtUtil.extractUserId(token));
         String role = JwtUtil.extractUserRole(token);
 
-//        if (!role.equalsIgnoreCase("admin") && !role.equalsIgnoreCase("manager")) {
-//            throw new RuntimeException("شما مجوز لازم برای انجام این عملیات را ندارید");
-//        }
+        Customer customer = customerRepository.findByUserId_Id(userId)
+                .orElseThrow(() -> new RuntimeException("خریدار پیدا نشد"));
 
-        if(!ordersRepository.existsById(orderId)) {
+        Optional<Orders> order = ordersRepository.findByCustomerIdAndOrderStatus(customer, OrderStatus.pending);
+
+        Orders orders = ordersRepository.findById(order.get().getOrderId()).orElseThrow(() -> new RuntimeException("سفارش مورد نظر پیدا نشد!"));
+
+        if (order.isEmpty()) {
+            throw new RuntimeException("هیچ سفارشی با وضعیت PENDING برای این خریدار یافت نشد");
+        }
+
+        if(!ordersRepository.existsById(order.get().getOrderId())) {
             throw new IllegalArgumentException("سفارش مورد نظر یافت نشد");
         }
 
-        if(!productVariantsRepository.existsById(variantId)) {
-            throw new IllegalArgumentException("کالای مورد نظر یافت نشد");
+        if(!orders.getOrderStatus().equals(OrderStatus.pending)){
+            throw new IllegalArgumentException("شما اجازه ویرایش اطلاعات این سفارش را ندارید");
         }
 
-        if(!orderItemRepository.existsById(orderItemId)) {
-            throw new IllegalArgumentException("ایتم خرید مورد نظر یافت نشد");
-        }
+        OrderItem exisit = orderItemRepository.findById(orderItemId).orElseThrow(() -> new RuntimeException("ایتم سفارش مورد نظر پیدا نشد!"));
 
-        Orders orders = ordersRepository.findById(orderId)
-                .orElseThrow(() -> new RuntimeException("سفارش پیدا نشد"));
-
-        ProductVariants products = productVariantsRepository.findById(variantId)
-                .orElseThrow(() -> new RuntimeException("محصول پیدا نشد"));
-
-        OrderItem update = orderItemRepository.findById(orderItemId)
-                .orElseThrow(() -> new RuntimeException("ایتم خرید مورد نظر یافت نشد"));
-
-        if(updatedOrderItem.getQuantity() == 0){
+        if(updatedOrderItem.getQuantity() == 0) {
             throw new IllegalArgumentException("تعداد کالا نباید صفر باشد");
         }
 
-        if(updatedOrderItem.getUnitPrice() == null){
-            throw new IllegalArgumentException("قیمت کالا نباید خالی باشد");
+        Integer stock = inventoryStockRepository.getTotalStockByVariantId(exisit.getVariantId().getVariantId());
+
+        if(stock < updatedOrderItem.getQuantity()){
+            throw new IllegalArgumentException("موجودی ناکافی است !");
         }
 
-        if(updatedOrderItem.getSubTotal() == null){
-            throw new IllegalArgumentException("مجموع قیمت کالا نباید خالی باشد");
-        }
-
-        if(updatedOrderItem.getTotalAmount() == null){
-            throw new IllegalArgumentException("قیمت نهایی نباید خالی باشد");
-        }
-
-
-        update.setOrderId(orders);
-        update.setVariantId(products);
-        update.setQuantity(updatedOrderItem.getQuantity());
-        update.setUnitPrice(updatedOrderItem.getUnitPrice());
-        update.setTotalAmount(updatedOrderItem.getTotalAmount());
-        update.setTaxAmount(updatedOrderItem.getTaxAmount());
-        update.setDiscountAmount(updatedOrderItem.getDiscountAmount());
-        update.setSubTotal(updatedOrderItem.getSubTotal());
-
-        return orderItemRepository.save(update);
+        exisit.setQuantity(updatedOrderItem.getQuantity());
+        return orderItemRepository.save(exisit);
     }
 
     public OrderItem getOrderItemById(UUID orderItemId) {
@@ -160,14 +158,19 @@ public class OrderItemService {
         UUID userId = UUID.fromString(JwtUtil.extractUserId(token));
         String role = JwtUtil.extractUserRole(token);
 
-//        if (!role.equalsIgnoreCase("admin") && !role.equalsIgnoreCase("staff")) {
-//            throw new RuntimeException("شما مجوز لازم برای انجام این عملیات را ندارید");
-//        }
 
         if (!orderItemRepository.existsById(orderItemId)) {
             throw new IllegalArgumentException("ایتم سفارش مورد نظر یافت نشد.");
         }
+
         OrderItem delete = orderItemRepository.findById(orderItemId).orElseThrow((() -> new EntityNotFoundException("سفارش مورد نظر پیدا نشد !")));
+
+        Orders orders = ordersRepository.findById(delete.getOrderId().getOrderId()).orElseThrow(() -> new RuntimeException("سفارش مورد نظر پیدا نشد!"));
+
+        if(!orders.getOrderStatus().equals(OrderStatus.pending)){
+            throw new IllegalArgumentException("شما اجازه حذف ایتم از سفارشی که در سبد خرید شما نیست را ندارید!");
+        }
+
         orderItemRepository.delete(delete);
     }
